+++ lib/inappstory_plugin.dart
@@ -16,6 +16,11 @@ export 'src/generated/pigeon_generated.g.dart'
         InAppStoryManagerHostApi,
         AppearanceManagerHostApi,
         GameReaderCallbackFlutterApi,
+        IASSingleStoryHostApi,
+        IASInAppMessagesHostApi,
+        IASOnboardingsHostApi,
+        IShowStoryCallbackFlutterApi,
+        OnboardingLoadCallbackFlutterApi,
         CallToActionCallbackFlutterApi;
 export 'src/widgets/decorators/decorators.dart';
 export 'src/widgets/placeholders/placeholders.dart';