+++ lib/src/callbacks/callbacks.dart
@@ -7,4 +7,6 @@ export 'ias_checkout_callback_mixin.dart';
 export 'ias_game_reader_callback_mixin.dart';
 export 'ias_good_item_selected_callback.dart';
 export 'ias_in_app_message_callback_mixin.dart';
+export 'ias_onboardings_callback_mixin.dart';
+export 'ias_show_story_callback_mixin.dart';
 export 'ias_skus_callback_impl.dart';