+++ lib/src/controllers/ias_manager.dart
@@ -1,4 +1,8 @@
+import 'dart:async';
+import 'dart:developer';
 import 'dart:ui';
+
+import 'package:async/async.dart';
 
 import '../callbacks/callbacks.dart'
     show GoodsCallbackFlutterApiImpl, SkusCallbackImpl;
@@ -6,7 +10,13 @@ import '../callbacks/ias_checkout_callback_impl.dart';
 import '../generated/checkout_generated.g.dart'
     show CheckoutManagerCallbackFlutterApi;
 import '../generated/pigeon_generated.g.dart'
-    show InAppStoryManagerHostApi, SkusCallbackFlutterApi;
+    show
+        InAppStoryManagerHostApi,
+        SkusCallbackFlutterApi,
+        IASInAppMessagesHostApi,
+        IASSingleStoryHostApi,
+        IASOnboardingsHostApi;
+import '../helpers/id_gen.dart';
 import 'logger.dart';
 
 class InAppStoryManager {
@@ -19,6 +29,9 @@ class InAppStoryManager {
 
   final _callbackImpl = GoodsCallbackFlutterApiImpl();
   final _checkoutCallbackImpl = IASCheckoutManagerCallbackImpl();
+  final _iam = IASInAppMessagesHostApi();
+  final _onboardings = IASOnboardingsHostApi();
+  final _singleStoryApi = IASSingleStoryHostApi();
 
   static final instance = InAppStoryManager._private();
 
@@ -102,5 +115,97 @@ class InAppStoryManager {
 
   Future<void> setOptions(Map<String, String> options) async {
     await _iasManager.setOptionKeys(options);
+  }
+
+  Future<bool> preloadInAppMessages({List<String>? ids}) async {
+    return await _iam.preloadMessages(ids: ids);
+  }
+
+  CancelableOperation<bool> showOnboarding(int limit,
+      {String? feed, List<String>? tags}) {
+    final uniqueId = idGenerator();
+    var operation = CancelableOperation.fromFuture(
+      _onboardings
+          .show(
+              limit: limit,
+              token: uniqueId,
+              feed: feed ?? 'onboarding',
+              tags: tags ?? <String>[])
+          .then((value) => true)
+          .catchError((error) {
+        log('[InAppStory]: showIAMById finished with error');
+        return false;
+      }),
+      onCancel: () async {
+        return _iam.cancelByToken(token: uniqueId);
+      },
+    );
+    return operation;
+  }
+
+  CancelableOperation<bool> showIAMById(String id,
+      {bool onlyPreloaded = false}) {
+    final uniqueId = idGenerator();
+    var operation = CancelableOperation.fromFuture(
+      _iam
+          .showById(id, uniqueId, onlyPreloaded: onlyPreloaded)
+          .then((value) => true)
+          .catchError((error) {
+        log('[InAppStory]: showIAMById finished with error');
+        return false;
+      }),
+      onCancel: () async {
+        return _iam.cancelByToken(token: uniqueId);
+      },
+    );
+    return operation;
+  }
+
+  CancelableOperation<bool> showIAMByEvent(String id) {
+    final uniqueId = idGenerator();
+    var operation = CancelableOperation.fromFuture(
+      _iam.showByEvent(id, uniqueId).then((value) => true).catchError((error) {
+        log('[InAppStory]: showIAMByEvent finished with error');
+        return false;
+      }),
+      onCancel: () async {
+        return _iam.cancelByToken(token: uniqueId);
+      },
+    );
+    return operation;
+  }
+
+  CancelableOperation<bool> showStory(String id) {
+    final uniqueId = idGenerator();
+    var operation = CancelableOperation.fromFuture(
+      _singleStoryApi
+          .show(storyId: id, token: uniqueId)
+          .then((value) => true)
+          .catchError((error) {
+        log('[InAppStory]: ShowStory finished with error');
+        return false;
+      }),
+      onCancel: () async {
+        return _singleStoryApi.cancelByToken(token: uniqueId);
+      },
+    );
+    return operation;
+  }
+
+  CancelableOperation<bool> showStoryOnce(String id) {
+    final uniqueId = idGenerator();
+    var operation = CancelableOperation.fromFuture(
+      _singleStoryApi
+          .showOnce(storyId: id, token: uniqueId)
+          .then((value) => true)
+          .catchError((error) {
+        log('[InAppStory]: showStoryOnce finished with error');
+        return false;
+      }),
+      onCancel: () async {
+        return _singleStoryApi.cancelByToken(token: uniqueId);
+      },
+    );
+    return operation;
   }
 }