+++ lib/src/data/enum.dart
@@ -3,3 +3,10 @@ enum CacheSize {
   medium,
   large,
 }
+
+enum IAMStatus {
+  running,
+  loading,
+  cancelled,
+  complete,
+}